+++ lib/menu-favorite.cpp
@@ -26,7 +26,7 @@ void MenuFavorite::flush(const AppVec &apps)
 {
     std::set<int32_t> app_set;
 
-    for (int i = 0; i < apps.size(); ++i)
+    for (int i = 0; i < (int)apps.size(); ++i)
     {
         auto &app = apps[i];
         auto &desktop_id = app->get_desktop_id();
